requirements.txt
blessed==1.21.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.7
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.2.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rsa==4.9.1
tqdm==4.66.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
